+++ pkg/querier/queryrange/queryrangebase/instrumentation.go
@@ -39,10 +39,10 @@ type InstrumentMiddlewareMetrics struct {
 }
 
 // NewInstrumentMiddlewareMetrics makes a new InstrumentMiddlewareMetrics.
-func NewInstrumentMiddlewareMetrics(registerer prometheus.Registerer) *InstrumentMiddlewareMetrics {
+func NewInstrumentMiddlewareMetrics(registerer prometheus.Registerer, metricsNamespace string) *InstrumentMiddlewareMetrics {
 	return &InstrumentMiddlewareMetrics{
 		duration: promauto.With(registerer).NewHistogramVec(prometheus.HistogramOpts{
-			Namespace: "cortex",
+			Namespace: metricsNamespace,
 			Name:      "frontend_query_range_duration_seconds",
 			Help:      "Total time spent in seconds doing query range requests.",
 			Buckets:   prometheus.DefBuckets,