+++ pkg/storage/hack/main.go
@@ -65,7 +65,7 @@ func getStore(cm storage.ClientMetrics) (storage.Store, *config.SchemaConfig, er
 		},
 	}
 
-	store, err := storage.NewStore(storeConfig, config.ChunkStoreConfig{}, schemaCfg, &validation.Overrides{}, cm, prometheus.DefaultRegisterer, util_log.Logger)
+	store, err := storage.NewStore(storeConfig, config.ChunkStoreConfig{}, schemaCfg, &validation.Overrides{}, cm, prometheus.DefaultRegisterer, util_log.Logger, "cortex")
 	return store, &schemaCfg, err
 }